removed old topical page and cleaned up after
