--- packages/app/src/domain/topical/common.ts
@@ -1,5 +1,0 @@
-import { colors } from '@corona-dashboard/common';
-
-export const COLOR_HAS_ONE_SHOT = colors.data.scale.blue[0];
-export const COLOR_FULLY_VACCINATED = colors.data.primary;
-export const COLOR_FULLY_BOOSTERED = colors.data.scale.blue[5];
