--- packages/app/src/domain/topical/common/categories.ts
@@ -1,12 +1,0 @@
-export const articleCategory = [
-  '__alles',
-  'vaccinaties',
-  'besmettingen',
-  'sterfte',
-  'ziekenhuizen',
-  'kwetsbare_groepen',
-  'vroege_indicatoren',
-  'gedrag',
-] as const;
-
-export type ArticleCategoryType = typeof articleCategory[number];
