--- packages/app/src/domain/topical/components/measurement-tile/index.ts
@@ -1,1 +1,0 @@
-export * from './measurement-tile';
