--- packages/app/src/domain/topical/components/subjects-list/index.ts
@@ -1,1 +1,0 @@
-export * from './subjects-list';
