+++ packages/app/src/domain/topical/components/topical-section-header.tsx
