--- packages/app/src/domain/topical/components/topical-tile/index.ts
@@ -1,1 +1,0 @@
-export * from './topical-tile';
