+++ packages/app/src/domain/topical/index.ts
@@ -1,9 +1,8 @@
-export * from './mini-tile-selector-layout';
-export { ArticleList } from './article-list';
-export { HighlightsTile } from './highlights-tile';
-export { ChoroplethTwoColumnLayout } from './choropleth-two-column-layout';
-export { MiniTileLayout } from './mini-tile-layout';
-export { MiniTrendTile } from './mini-trend-tile';
-export { MiniVaccinationCoverageTile } from './mini-vaccination-coverage-tile';
-export { TopicalSectionHeader } from './topical-section-header';
-export { VaccinationCoverageChoropleth } from './vaccination-coverage-choropleth';
+export *  from './components/topical-tile';
+export *  from './components/topical-header';
+export *  from './components/theme-header';
+export *  from './components/subjects-list';
+export *  from './components/measurement-tile';
+export *  from './components/search';
+export { ArticleList } from './components/article-list';
+export { TopicalSectionHeader } from './components/topical-section-header';
