--- packages/app/src/domain/topical/mini-vaccination-coverage-tile.tsx
@@ -1,89 +1,0 @@
-import { Box } from '~/components/base';
-import { Markdown } from '~/components/markdown';
-import {
-  COLOR_FULLY_VACCINATED,
-  COLOR_HAS_ONE_SHOT,
-  COLOR_FULLY_BOOSTERED,
-} from './common';
-import { Bar } from '../vaccine/components/bar';
-import { MiniTile, MiniTileProps } from './mini-tile';
-
-type MiniVaccinationCoverageTileProps = {
-  oneShotPercentage: number | null;
-  fullyVaccinatedPercentage: number | null;
-  boosterShotAdministered?: number | null;
-  oneShotPercentageLabel?: string | null;
-  fullyVaccinatedPercentageLabel?: string | null;
-  boosterShotAdministeredLabel?: string | null;
-  oneShotBarLabel: string;
-  fullyVaccinatedBarLabel: string;
-  boosterShotAdministeredBarLabel?: string;
-} & Omit<MiniTileProps, 'children'>;
-
-export function MiniVaccinationCoverageTile(
-  props: MiniVaccinationCoverageTileProps
-) {
-  const {
-    oneShotPercentage,
-    fullyVaccinatedPercentage,
-    boosterShotAdministered,
-    oneShotPercentageLabel,
-    fullyVaccinatedPercentageLabel,
-    boosterShotAdministeredLabel,
-    oneShotBarLabel,
-    fullyVaccinatedBarLabel,
-    boosterShotAdministeredBarLabel,
-    ...tileProps
-  } = props;
-
-  return (
-    <MiniTile {...tileProps}>
-      <Box display="flex" flexDirection="column" spacing={3}>
-        {boosterShotAdministered && boosterShotAdministeredBarLabel && (
-          <LabeledBar
-            value={boosterShotAdministered}
-            color={COLOR_FULLY_BOOSTERED}
-            valueLabel={boosterShotAdministeredLabel}
-            barLabel={boosterShotAdministeredBarLabel}
-          />
-        )}
-        <LabeledBar
-          value={fullyVaccinatedPercentage}
-          color={COLOR_FULLY_VACCINATED}
-          valueLabel={fullyVaccinatedPercentageLabel}
-          barLabel={fullyVaccinatedBarLabel}
-        />
-        <LabeledBar
-          value={oneShotPercentage}
-          color={COLOR_HAS_ONE_SHOT}
-          valueLabel={oneShotPercentageLabel}
-          barLabel={oneShotBarLabel}
-        />
-      </Box>
-    </MiniTile>
-  );
-}
-
-type LabeledBarProps = {
-  value: number | null;
-  valueLabel?: string | null;
-  color: string;
-  barLabel: string;
-};
-
-function LabeledBar(props: LabeledBarProps) {
-  const { value, valueLabel, color, barLabel } = props;
-
-  return (
-    <Box spacing={1}>
-      <Markdown content={barLabel} />
-      <Bar
-        value={value}
-        color={color}
-        label={valueLabel}
-        height={20}
-        showAxisValues
-      />
-    </Box>
-  );
-}
